Make functions open so they can be subclasses (#2)
When trying to implement a custom `URLSessionTask` I ran into the compiler error that they were not `open` so we could override them or provide our own implementation.

Taking more notes from the `swift-corelibs-foundation` repo, it seems that most of these functions should be `open` so that they could be subclassed if needed.