Make state read only like it should be
We need to ensure that we're matching the semantics of the URLSession
APIs otherwise things that are built on top of it might fail to compile
since the requirements are different. In this case we had someone
building on top of URLSession semantics that state was read only. They
had a protocol which also outlined state as read only which failed to
compile due to our changes making it read/write.